change ints to uint8 and uint16

+++ src/bb01/bb01.ino
@@ -17,7 +17,7 @@ void setup() {
  */
 void loop() {
 
-    int buttonClicked;
+    uint8_t buttonClicked;
 
     // poll buttons to see if one or more is pressed
     buttonClicked = buttons->poll();
